added the refactor for PPU addr
added the refactor for ppu addr
